+++ docs/html/search/all_4.js
@@ -1,0 +1,6 @@
+var searchData=
+[
+  ['debug',['debug',['../classtacopie_1_1logger__iface.html#a156abb02ab852ea4033fc13f4902ee7a',1,'tacopie::logger_iface::debug()'],['../classtacopie_1_1logger.html#aff31bbc7d3fdbbe60a2331fe24ec76ff',1,'tacopie::logger::debug(const std::string &amp;msg, const std::string &amp;file, std::size_t line)'],['../classtacopie_1_1logger.html#ae7dd235972bbf86a017fc39b3af80efeaad42f6697b035b7580e4fef93be20b4d',1,'tacopie::logger::debug()'],['../namespacetacopie.html#a0833874b07ad56b5ba2c7847f1e514b0',1,'tacopie::debug()']]],
+  ['disconnect',['disconnect',['../classtacopie_1_1tcp__client.html#a7562e0bfa24912595d6f695f848b9e51',1,'tacopie::tcp_client']]],
+  ['disconnection_5fhandler_5ft',['disconnection_handler_t',['../classtacopie_1_1tcp__client.html#aca5df52e5ee6fa673cf212532ada1453',1,'tacopie::tcp_client']]]
+];