+++ docs/html/search/all_5.js
@@ -1,0 +1,6 @@
+var searchData=
+[
+  ['error',['error',['../classtacopie_1_1logger__iface.html#a18f9c02fc19be4b9900ac9fb1a361624',1,'tacopie::logger_iface::error()'],['../classtacopie_1_1logger.html#a3fe1be02ac2f4e4fe44a0bdaf8359546',1,'tacopie::logger::error(const std::string &amp;msg, const std::string &amp;file, std::size_t line)'],['../classtacopie_1_1logger.html#ae7dd235972bbf86a017fc39b3af80efeacb5e100e5a9a3e7f6d1fd97512215282',1,'tacopie::logger::error()'],['../namespacetacopie.html#a48f7b277b1559f2333c1186fa629ef01',1,'tacopie::error()']]],
+  ['error_2ehpp',['error.hpp',['../error_8hpp.html',1,'']]],
+  ['event_5fcallback_5ft',['event_callback_t',['../classtacopie_1_1io__service.html#abb66850c32d9c724f4418d77bd04bcfd',1,'tacopie::io_service']]]
+];