+++ docs/html/search/all_d.js
@@ -1,0 +1,9 @@
+var searchData=
+[
+  ['poll',['poll',['../classtacopie_1_1io__service.html#a03813e0e91d70d367f9fb0ff0d86d3dd',1,'tacopie::io_service']]],
+  ['process_5fevents',['process_events',['../classtacopie_1_1io__service.html#a90a2ab7b6bbaf10bbc0607ce371cc806',1,'tacopie::io_service']]],
+  ['process_5frd_5fevent',['process_rd_event',['../classtacopie_1_1io__service.html#a65c3ef9fa1473708b81b05029f6861f9',1,'tacopie::io_service']]],
+  ['process_5fread',['process_read',['../classtacopie_1_1tcp__client.html#a40308b6a1642a79a35a2b2d9c9b9a3f7',1,'tacopie::tcp_client']]],
+  ['process_5fwr_5fevent',['process_wr_event',['../classtacopie_1_1io__service.html#a55e334cdf75187b9296c03447e32a63f',1,'tacopie::io_service']]],
+  ['process_5fwrite',['process_write',['../classtacopie_1_1tcp__client.html#afd5e43d5f44930894de10ce05268607f',1,'tacopie::tcp_client']]]
+];