+++ docs/html/search/classes_3.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['self_5fpipe',['self_pipe',['../classtacopie_1_1self__pipe.html',1,'tacopie']]]
+];