+++ sources/network/windows/tcp_socket.cpp
@@ -20,10 +20,10 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-#include <tacopie/error.hpp>
-#include <tacopie/logger.hpp>
 #include <tacopie/network/tcp_server.hpp>
-#include <tacopie/typedefs.hpp>
+#include <tacopie/utils/error.hpp>
+#include <tacopie/utils/logger.hpp>
+#include <tacopie/utils/typedefs.hpp>
 
 #include <cstring>
 
@@ -32,78 +32,12 @@
 
 namespace tacopie {
 
-//!
-//! ctor & dtor
-//!
-
-tcp_socket::tcp_socket(void)
-: m_fd(__TACOPIE_INVALID_FD)
-, m_host("")
-, m_port(0)
-, m_type(type::UNKNOWN) { __TACOPIE_LOG(debug, "create tcp_socket"); }
-
-//!
-//! custom ctor
-//! build socket from existing file descriptor
-//!
-
-tcp_socket::tcp_socket(fd_t fd, const std::string& host, std::uint32_t port, type t)
-: m_fd(fd)
-, m_host(host)
-, m_port(port)
-, m_type(t) { __TACOPIE_LOG(debug, "create tcp_socket"); }
-
-//!
-//! Move constructor
-//!
-
-tcp_socket::tcp_socket(tcp_socket&& socket)
-: m_fd(std::move(socket.m_fd))
-, m_host(socket.m_host)
-, m_port(socket.m_port)
-, m_type(socket.m_type) {
-  socket.m_fd   = __TACOPIE_INVALID_FD;
-  socket.m_type = type::UNKNOWN;
-
-  __TACOPIE_LOG(debug, "moved tcp_socket");
-}
-
-//!
-//! client socket operations
-//!
-
-std::vector<char>
-tcp_socket::recv(std::size_t size_to_read) {
-  create_socket_if_necessary();
-  check_or_set_type(type::CLIENT);
-
-  std::vector<char> data(size_to_read, 0);
-
-  ssize_t rd_size = ::recv(m_fd, const_cast<char*>(data.data()), size_to_read, 0);
-
-  if (rd_size == SOCKET_ERROR) { __TACOPIE_THROW(error, "recv() failure"); }
-
-  if (rd_size == 0) { __TACOPIE_THROW(warn, "nothing to read, socket has been closed by remote host"); }
-
-  data.resize(rd_size);
-
-  return data;
-}
-
-std::size_t
-tcp_socket::send(const std::vector<char>& data, std::size_t size_to_write) {
-  create_socket_if_necessary();
-  check_or_set_type(type::CLIENT);
-
-  ssize_t wr_size = ::send(m_fd, data.data(), size_to_write, 0);
-
-  if (wr_size == SOCKET_ERROR) { __TACOPIE_THROW(error, "send() failure"); }
-
-  return wr_size;
-}
-
 void
-tcp_socket::connect(const std::string& host, std::uint32_t port) {
+tcp_socket::connect(const std::string& host, std::uint32_t port, std::uint32_t timeout_msecs) {
+  //! Reset host and port
+  m_host = host;
+  m_port = port;
+
   create_socket_if_necessary();
   check_or_set_type(type::CLIENT);
 
@@ -124,7 +58,49 @@ tcp_socket::connect(const std::string& host, std::uint32_t port) {
 
   freeaddrinfo(result);
 
-  if (::connect(m_fd, (const struct sockaddr*) &server_addr, sizeof(server_addr)) == -1) { __TACOPIE_THROW(error, "connect() failure"); }
+  if (timeout_msecs > 0) {
+    //! for timeout connection handling:
+    //!  1. set socket to non blocking
+    //!  2. connect
+    //!  3. poll select
+    //!  4. check connection status
+    u_long mode = 1;
+    if (ioctlsocket(m_fd, FIONBIO, &mode) != 0) {
+      close();
+      __TACOPIE_THROW(error, "connect() set non-blocking failure");
+    }
+  }
+
+  int ret = ::connect(m_fd, (const struct sockaddr*) &server_addr, sizeof(server_addr));
+  if (ret == -1 && WSAGetLastError() != WSAEWOULDBLOCK) {
+    close();
+    __TACOPIE_THROW(error, "connect() failure");
+  }
+
+  if (timeout_msecs > 0) {
+    timeval tv;
+    tv.tv_sec  = (timeout_msecs / 1000);
+    tv.tv_usec = ((timeout_msecs - (tv.tv_sec * 1000)) * 1000);
+
+    FD_SET set;
+    FD_ZERO(&set);
+    FD_SET(m_fd, &set);
+
+    //! 1 means we are connected.
+    //! 0 means a timeout.
+    if (select(0, NULL, &set, NULL, &tv) == 1) {
+      //! Set back to blocking mode as the user of this class is expecting
+      u_long mode = 0;
+      if (ioctlsocket(m_fd, FIONBIO, &mode) != 0) {
+        close();
+        __TACOPIE_THROW(error, "connect() set blocking failure");
+      }
+    }
+    else {
+      close();
+      __TACOPIE_THROW(error, "connect() timed out");
+    }
+  }
 }
 
 //!
@@ -133,6 +109,10 @@ tcp_socket::connect(const std::string& host, std::uint32_t port) {
 
 void
 tcp_socket::bind(const std::string& host, std::uint32_t port) {
+  //! Reset host and port
+  m_host = host;
+  m_port = port;
+
   create_socket_if_necessary();
   check_or_set_type(type::SERVER);
 
@@ -153,29 +133,6 @@ tcp_socket::bind(const std::string& host, std::uint32_t port) {
   if (::bind(m_fd, (const struct sockaddr*) &server_addr, sizeof(server_addr)) == SOCKET_ERROR) { __TACOPIE_THROW(error, "bind() failure"); }
 }
 
-void
-tcp_socket::listen(std::size_t max_connection_queue) {
-  create_socket_if_necessary();
-  check_or_set_type(type::SERVER);
-
-  if (::listen(m_fd, max_connection_queue) == SOCKET_ERROR) { __TACOPIE_THROW(debug, "listen() failure"); }
-}
-
-tcp_socket
-tcp_socket::accept(void) {
-  create_socket_if_necessary();
-  check_or_set_type(type::SERVER);
-
-  struct sockaddr_in client_info;
-  socklen_t client_info_struct_size = sizeof(client_info);
-
-  fd_t client_fd = ::accept(m_fd, (struct sockaddr*) &client_info, &client_info_struct_size);
-
-  if (client_fd == __TACOPIE_INVALID_FD) { __TACOPIE_THROW(error, "accept() failure"); }
-
-  return {client_fd, inet_ntoa(client_info.sin_addr), client_info.sin_port, type::CLIENT};
-}
-
 //!
 //! general socket operations
 //!
@@ -189,88 +146,6 @@ tcp_socket::close(void) {
 
   m_fd   = __TACOPIE_INVALID_FD;
   m_type = type::UNKNOWN;
-}
-
-//!
-//! create a new socket if no socket has been initialized yet
-//!
-
-void
-tcp_socket::create_socket_if_necessary(void) {
-  if (m_fd != __TACOPIE_INVALID_FD) { return; }
-
-  //! new TCP socket
-  m_fd   = socket(AF_INET, SOCK_STREAM, 0);
-  m_type = type::UNKNOWN;
-
-  if (m_fd == __TACOPIE_INVALID_FD) { __TACOPIE_THROW(error, "tcp_socket::create_socket_if_necessary: socket() failure"); }
-}
-
-//!
-//! check whether the current socket has an approriate type for that kind of operation
-//! if current type is UNKNOWN, update internal type with given type
-//!
-
-void
-tcp_socket::check_or_set_type(type t) {
-  if (m_type != type::UNKNOWN && m_type != t) { __TACOPIE_THROW(error, "trying to perform invalid operation on socket"); }
-
-  m_type = t;
-}
-
-//!
-//! get socket name information
-//!
-
-const std::string&
-tcp_socket::get_host(void) const {
-  return m_host;
-}
-
-std::uint32_t
-tcp_socket::get_port(void) const {
-  return m_port;
-}
-
-//!
-//! get socket type
-//!
-
-tcp_socket::type
-tcp_socket::get_type(void) const {
-  return m_type;
-}
-
-//!
-//! set type, should be used if some operations determining socket type
-//! have been done on the behalf of the tcp_socket instance
-//!
-
-void
-tcp_socket::set_type(type t) {
-  m_type = t;
-}
-
-//!
-//! direct access to the underlying fd
-//!
-
-fd_t
-tcp_socket::get_fd(void) const {
-  return m_fd;
-}
-
-//!
-//! comparison operator
-//!
-bool
-tcp_socket::operator==(const tcp_socket& rhs) const {
-  return m_fd == rhs.m_fd && m_type == rhs.m_type;
-}
-
-bool
-tcp_socket::operator!=(const tcp_socket& rhs) const {
-  return !operator==(rhs);
 }
 
 } // namespace tacopie